Start passing down a 'speculation id' to deoptimizations.     Use it to record GuardNode id during guard lowering as a debug feature for now

+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.graal.api.meta/src/com/oracle/graal/api/meta/MetaAccessProvider.java
@@ -74,7 +74,7 @@ public interface MetaAccessProvider {
      * 
      * @return the encoded value as an integer
      */
-    Constant encodeDeoptActionAndReason(DeoptimizationAction action, DeoptimizationReason reason, int speculationId);
+    Constant encodeDeoptActionAndReason(DeoptimizationAction action, DeoptimizationReason reason, short speculationId);
 
     DeoptimizationReason decodeDeoptReason(Constant constant);
